Switch project compiler back to MSVC, fixed resulting compilation errors, made transparent material rendering/ordering work
The goal of this compiler switchback is to allow C++20/23 easily in the future:
- Move third-party libraries to external #includes in Project Settings, so we can turn off their warnings "beyond our control" (all enforced as errors in the project)
- Fixed all resulting compilation errors

While repairing the above, noticed the transparence was not correctly working per shader. In fact, only a single alpha value (the one of the last Entity) was applied for all Entities of the simulation, overwriting the other ones:
- Added per-frame update of the 'alpha' material parameter for all Entities using a shader that makes use of it (i.e. in GLSL code, Celestial Body with alpha = 0; Orbit + Rings for alpha != 0)
- Added validity checks when loading material model parameters from .mtl files. Fixed procesing of 'Dissolve' (= alpha) parameter. Updated material parameter comments in .mtl files
- Renamed corresponding GLSL parameter to 'fu_Alpha' across the board
- Sorted out correctly per distance from farthest to closest to camera all Scene Entity with level of transparency
- Removed some dead code, brought some up to standards, and did renaming pass

+++ OpenGLProject/Components/Meshes/MeshComponent.cpp
@@ -74,11 +74,11 @@ void MeshComponent::Render(const unsigned int mode) const
 			std::cout << "ERROR::MESH - Any call to this method should have a non-null IBO!" << std::endl;
 		}
 
-		Renderer::Draw(mode, indices.size(), nullptr);
+		Renderer::Draw(mode, static_cast<int32_t>(indices.size()), nullptr);
 	}
 	else
 	{
-		Renderer::Draw(mode, 0, vertices.size());
+		Renderer::Draw(mode, 0, static_cast<int32_t>(vertices.size()));
 	}
 
 	vao->Unbind();
